requirements.txt
Flask==3.1.*
PyPDF2>=3.0
python-docx>=1.0
scikit-learn>=1.4
nltk>=3.9
openai-whisper>=20231117
pymongo>=4.6
sentence-transformers>=2.5
Werkzeug>=3.0
Flask-Session>=0.5
google-generativeai>=0.8
python-dotenv>=1.0.1
cloudinary>=1.40.0
gunicorn>=21.2.0
requests>=2.31
